+++ packages/apollo-env/src/index.ts
@@ -2,3 +2,4 @@ import "./polyfills";
 
 export * from "./typescript-utility-types";
 export * from "../lib/fetch";
+export * from "./utils";