+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - apollo
   - JSON flag for service:check output [#1079](https://github.com/apollographql/apollo-tooling/pull/1079)
+- apollo-env
+  - Add environment-aware createHash function to apollo-env [#1110](https://github.com/apollographql/apollo-tooling/pull/1110)
 
 ## `apollo@2.6.0`, `apollo-env@0.3.4`, `apollo-language-server@1.5.3`, `apollo-codegen-flow@0.32.9`, `apollo-codegen-scala@0.33.5`, `apollo-codegen-swift@0.32.9`, `apollo-codegen-typescript@0.32.10`, `apollo-graphql@0.1.2`